requirements.txt
flask
python-dotenv
coverage==4.5.2
pytest
pytest-cov==2.6.1
pytest-flakes==2.0.0
pytest-pep8
pytest-pythonpath
requests
pytest-html
dnspython
pymongo

pytest-mongodb
mongomock
pyyaml>=5.1
pytest>=2.5.2
mock
flake8>=2.1.0
Flask-PyMongo